Move backend exception into infra module
