+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/exception/TableLockWaitTimeoutException.java
@@ -15,20 +15,24 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.proxy.backend.exception;
+package org.apache.shardingsphere.infra.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.infra.util.exception.ShardingSphereInsideException;
 
 /**
- * DB create exists exception.
+ * Table lock wait timeout exception.
  */
 @RequiredArgsConstructor
 @Getter
-public final class DBCreateExistsException extends ShardingSphereInsideException {
+public final class TableLockWaitTimeoutException extends ShardingSphereInsideException {
     
-    private static final long serialVersionUID = 779787160167652641L;
+    private static final long serialVersionUID = 2599713085782288003L;
     
-    private final String databaseName;
+    private final String schemaName;
+    
+    private final String tableName;
+    
+    private final Long timeoutMilliseconds;
 }
